safari-600.7-branch/LayoutTests/platform/ios-sim/storage/indexeddb/unblocked-version-changes-expected.txt

CONSOLE MESSAGE: line 12: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
Ensure that metadata remains correct when an aborted version change is followed by another.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

dbname = "unblocked-version-changes.html"
indexedDB.deleteDatabase(dbname)
FAIL indexedDB.deleteDatabase(dbname) threw exception TypeError: null is not an object (evaluating 'indexedDB.deleteDatabase')
FAIL Unexpected error: TypeError: undefined is not an object (evaluating 'request.onblocked = unexpectedBlockedCallback')
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
